api/ml/pyproject.toml
[project]
name = "attribution-system-ml-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask-cors>=5.0.1",
    "flask-restx>=1.3.0",
    "gensim>=4.3.3",
    "gunicorn>=23.0.0",
    "joblib>=1.4.2",
    "nameparser>=1.1.3",
    "numpy>=1.26.4",
    "pdfplumber>=0.11.5",
    "pip>=25.0.1",
    "scikit-learn==1.5.2",
    "sentence-transformers>=3.4.1",
    "spacy>=3.8.4",
    "torch>=2.6.0",
    "transformers>=4.49.0",
]
